optimize course dashboard

+++ client/src/pages/admin/CourseDashboard/CourseList/index.jsx
@@ -2,6 +2,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { getCourses, deleteCourse } from "../courseSlice";
 import Button from "components/Button";
 import style from "./CourseList.module.scss";
+import { useEffect } from "react";
 
 const CourseList = ({ setId }) => {
   const dispatch = useDispatch();
@@ -11,8 +12,16 @@ const CourseList = ({ setId }) => {
     dispatch(getCourses({ skip: courses.length, limit: 5 }));
   };
   const deleteHandle = (id) => {
-    dispatch(deleteCourse(id));
+    if (window.confirm("Are you sure to delete this course")) {
+      dispatch(deleteCourse(id));
+    }
   };
+
+  useEffect(() => {
+    if (!courses.length) {
+      dispatch(getCourses({ skip: 0, limit: 5 }));
+    }
+  }, [courses.length, dispatch]);
 
   return (
     <div className={style.courseList}>
@@ -24,7 +33,7 @@ const CourseList = ({ setId }) => {
             <th>Tuition</th>
             <th>
               <div className="d-flex justify-content-center">
-                <Button type="outline" onClick={() => setId(0)}>
+                <Button variant="outline" onClick={() => setId(0)}>
                   Add New Course
                 </Button>
               </div>
@@ -45,10 +54,10 @@ const CourseList = ({ setId }) => {
               </td>
               <td>
                 <div className="d-flex justify-content-center">
-                  <Button type="outline" onClick={() => setId(course._id)}>
+                  <Button variant="outline" onClick={() => setId(course._id)}>
                     Edit
                   </Button>
-                  <Button type="outline" onClick={() => deleteHandle(course._id)}>
+                  <Button variant="outline" onClick={() => deleteHandle(course._id)}>
                     Delete
                   </Button>
                 </div>
@@ -57,7 +66,9 @@ const CourseList = ({ setId }) => {
           ))}
         </tbody>
       </table>
-      <Button type="outline" className={style.loadButton} onClick={loadHandle}>Load more</Button>
+      <Button variant="outline" className={style.loadButton} onClick={loadHandle}>
+        Load more
+      </Button>
     </div>
   );
 };
